refactor: 세션, 통계 패키지 분리
