+++ src/main/kotlin/com/yapp/brake/session/infrastructure/jpa/session/SessionRepository.kt
@@ -1,4 +1,4 @@
-package com.yapp.brake.session.infrastructure.jpa
+package com.yapp.brake.session.infrastructure.jpa.session
 
 import org.springframework.data.jpa.repository.JpaRepository
