+++ src/main/kotlin/com/yapp/brake/session/infrastructure/jpa/statistics/DailySessionStatisticsId.kt
@@ -1,4 +1,4 @@
-package com.yapp.brake.session.infrastructure.jpa
+package com.yapp.brake.session.infrastructure.jpa.statistics
 
 import java.io.Serializable
 import java.time.LocalDate
