+++ src/main/kotlin/com/yapp/brake/session/infrastructure/jpa/statistics/DailySessionStatisticsJpaWriter.kt
@@ -1,4 +1,4 @@
-package com.yapp.brake.session.infrastructure.jpa
+package com.yapp.brake.session.infrastructure.jpa.statistics
 
 import com.yapp.brake.session.infrastructure.DailySessionStatisticsWriter
 import com.yapp.brake.session.model.DailySessionStatistics
