+++ src/main/kotlin/com/yapp/brake/session/infrastructure/jpa/statistics/DailySessionStatisticsRepository.kt
@@ -1,4 +1,4 @@
-package com.yapp.brake.session.infrastructure.jpa
+package com.yapp.brake.session.infrastructure.jpa.statistics
 
 import org.springframework.data.jpa.repository.JpaRepository
